Move constants under utils
